Played with player colliders
Made the top of the player curved, so we don't get stuck on things.
Moved the Death Message canvas up a bit, so we can clearly see the
message.
Dropped the temperature in the jungle to zero, now making it the frozen
jungle.
Fixed HUD not updating the showing the correct amount of lives when a
level restarts.
Spawnpoint now removes duplicate players.